Remove erroneous int cast
float should stay float,
as to be Util.stringify():ied

+++ src/main/chariot/chariot/internal/impl/BoardHandler.java
@@ -127,7 +127,7 @@ public class BoardHandler extends ChallengesAuthCommonImpl implements BoardAuth 
             @Override
             public SeekParams clock(float initialMinutes, int incrementSeconds) {
                 return builder
-                    .add("time", (int)initialMinutes)
+                    .add("time", initialMinutes)
                     .add("increment", incrementSeconds)
                     .proxy();
             }
